+++ app/source.ts
@@ -2,9 +2,6 @@ import { map } from '../.map';
 import { createMDXSource } from 'fumadocs-mdx';
 import { loader } from 'fumadocs-core/source';
 import { PageTree } from 'fumadocs-core/server';
-import { getEnvConfig } from './../configs/env/index.mjs';
-
-const envConfig = getEnvConfig();
 
 export const {
   getPage,
@@ -20,18 +17,23 @@ export const pageTree: PageTree.Root = {
   name: 'Nova UI',
   children: [
     { type: 'separator', name: 'Getting Started' },
-    { type: 'page', name: 'Setup', url: `${envConfig.SERVER.BASE_PATH}/docs` },
-    { type: 'page', name: 'Changelog', url: `${envConfig.SERVER.BASE_PATH}/docs/changelog` },
+    { type: 'page', name: 'Setup', url: '/docs' },
+    { type: 'page', name: 'Changelog', url: '/docs/changelog' },
 
     { type: 'separator', name: ' 按钮' },
-    { type: 'page', name: 'Button', url: `${envConfig.SERVER.BASE_PATH}/docs/button` },
+    { type: 'page', name: 'Button', url: '/docs/button' },
 
     { type: 'separator', name: '开关' },
-    { type: 'page', name: 'Switch', url: `${envConfig.SERVER.BASE_PATH}/docs/switch` },
+    { type: 'page', name: 'Switch', url: '/docs/switch' },
 
     { type: 'separator', name: '弹出层' },
-    { type: 'page', name: 'Dialog', url: `${envConfig.SERVER.BASE_PATH}/docs/dialog` },
-    { type: 'page', name: 'Drawer', url: `${envConfig.SERVER.BASE_PATH}/docs/drawer` },
-    { type: 'page', name: 'Credenza', url: `${envConfig.SERVER.BASE_PATH}/docs/credenza` },
+    { type: 'page', name: 'Dialog', url: '/docs/dialog' },
+    { type: 'page', name: 'Drawer', url: '/docs/drawer' },
+    { type: 'page', name: 'Credenza', url: '/docs/credenza' },
+    { type: 'page', name: 'Tooltip', url: '/docs/tooltip' },
+
+    { type: 'separator', name: '标签栏' },
+    { type: 'page', name: 'BlockTabs', url: '/docs/block-tabs' },
+    { type: 'page', name: 'TextTabs', url: '/docs/text-tabs' }
   ],
 };