+++ content/docs/changelog.md
@@ -3,4 +3,25 @@ title: Changelog
 description: Latest updates and announcements
 ---
 
+## 2025-02-25
+
+**New Component**
+
+* 🔥 新增 [`Tooltip`](/docs/tooltip) 组件
+* 🔥 新增 [`BlockTabs`](/docs/block-tabs) 组件
+* 🔥 新增 [`TextTabs`](/docs/text-tabs) 组件
+
+## 2025-02-22
+
+**Feature**
+
+* 🆕 `Dialog.Content` 支持 `showOverlay` 属性，控制展示遮罩层
+
 ## 2025-02-17
+
+**New Component**
+
+* 🔥 新增 [`Button`](/docs/button) 组件
+* 🔥 新增 [`Dialog`](/docs/dialog) 组件
+* 🔥 新增 [`Drawer`](/docs/drawer) 组件
+* 🔥 新增 [`Credenza`](/docs/drawer) 组件